managing addresses - generic listing components
