+++ packages/renderer/src/screens/Import/Addresses/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2025 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export * from './Address';
+export * from './Listing';
